// src/api/orval/model/ovpnFileConfigResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface OvpnFileConfigResponse {
  id?: number;
  vpnServerId?: number;
  /** @nullable */
  vpnServerIp?: string | null;
  vpnServerPort?: number;
  /** @nullable */
  configTemplate?: string | null;
}